pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parliament-of-owls"
version = "0.1.0"
description = "Query multiple LLMs and deep research APIs in parallel"
requires-python = ">=3.10"
dependencies = [
    "llm>=0.26",
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
owl = "owl.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
